O sistema possibilita a edição de lojas, bem como a exclusão de uma loja. As páginas e recursos inerentes a essas funcionalidades foram adicionadas. A mesmo tela de criação de uma loja é utlizado na edição, a alteração é feita de forma dinâmica, quando se trata de uma nova loja, a tela correspondente é apresentada, da mesma forma ocorre quando for uma mudança numa loja já existente.

+++ src/app/components/restaurant-detail/shopping-cart/shopping-cart.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ItemMenu } from '../item-menu/item-menu.model';
 import { ShoppingCartService } from './shopping-cart.service';
+import { ConfirmationDialogComponent } from 'app/components/shared/confirmation-dialog/confirmation-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'mt-shopping-cart',
@@ -8,7 +10,7 @@ import { ShoppingCartService } from './shopping-cart.service';
 })
 export class ShoppingCartComponent implements OnInit {
 
-  constructor(private shoppingCartService: ShoppingCartService) { }
+  constructor(private shoppingCartService: ShoppingCartService, private dialog: MatDialog) { }
 
   ngOnInit() {
   }
@@ -26,7 +28,22 @@ export class ShoppingCartComponent implements OnInit {
   }
 
   removeAllItems() {
-    return this.shoppingCartService.removeAllItems();
+
+    // janela de diálogo para confirmação da exclusão do registro
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '250px',
+      panelClass: 'my-custom-dialog',
+      data: { description: 'Quer mesmo remover todos os itens do carrinho?' }
+    });
+
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      // se sim, chama o serviço para remoção de todos os itens
+      if(result) {
+        return this.shoppingCartService.removeAllItems();
+      }
+    });
+
   }
 
   totalToPay(): number {
